refactor: consolidate stake options into single component
