+++ src/cli/index.ts
@@ -15,8 +15,6 @@ export * as pagination from './pagination'
 export * as passphrase from './passphrase'
 export * as privateKey from './privateKey'
 export * as stake from './stake'
-export * as stakeStatus from './stakeStatus'
-export * as stakeType from './stakeType'
 export * as verbose from './verbose'
 export * as wallet from './wallet'
 export * as yes from './yes'
